pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "notam-downloader"
version = "0.1.0"
description = "GCAA UAE NOTAM OMAE_ValidNOTAM PDF downloader and parser"
requires-python = ">=3.11"
dependencies = [
    "playwright",
    "pdfplumber",
    "aiosmtplib",
    "apscheduler",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[project.scripts]
notam = "notam.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/notam"]

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"

# ---------------------------------------------------------------------------
# ruff
# ---------------------------------------------------------------------------

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]  # line length enforced by formatter, not linter
